Add a lot of unittests ; add del command

+++ ergonomica/lib/lib/ergo_add.py
@@ -19,7 +19,7 @@ def add(argc):
     parsed_numbers = []
     for i in range(len(argc.args['NUMBERS'])):
         try:
-            parsed_numbers.append(int(argc.args['NUMBERS'][i]))
+            parsed_numbers.append(float(argc.args['NUMBERS'][i]))
         except ValueError:
             print("[ergo: add]: '{}' (index={}) not a number".format(argc.args['NUMBERS'][i], i))
